+++ glean-core/android/src/main/java/mozilla/telemetry/glean/Glean.kt
@@ -20,7 +20,6 @@ import mozilla.telemetry.glean.config.FfiConfiguration
 import mozilla.telemetry.glean.utils.getLocaleTag
 import java.io.File
 import mozilla.telemetry.glean.rust.LibGleanFFI
-import mozilla.telemetry.glean.rust.MetricHandle
 import mozilla.telemetry.glean.rust.getAndConsumeRustString
 import mozilla.telemetry.glean.rust.toBoolean
 import mozilla.telemetry.glean.rust.toByte
@@ -55,8 +54,7 @@ open class GleanInternalAPI internal constructor () {
         internal const val GLEAN_DATA_DIR: String = "glean_data"
     }
 
-    // `internal` so this can be modified for testing
-    internal var handle: MetricHandle = 0L
+    private var initialized: Boolean = false
 
     internal lateinit var configuration: Configuration
 
@@ -152,10 +150,10 @@ open class GleanInternalAPI internal constructor () {
             delayPingLifetimeIO = false
         )
 
-        handle = LibGleanFFI.INSTANCE.glean_initialize(cfg)
+        initialized = LibGleanFFI.INSTANCE.glean_initialize(cfg).toBoolean()
 
         // If initialization of Glean fails we bail out and don't initialize further.
-        if (handle == 0L) {
+        if (!initialized) {
             return
         }
 
@@ -166,7 +164,7 @@ open class GleanInternalAPI internal constructor () {
         // If this is the first time ever the Glean SDK runs, make sure to set
         // some initial core metrics in case we need to generate early pings.
         // The next times we start, we would have them around already.
-        val isFirstRun = LibGleanFFI.INSTANCE.glean_is_first_run(handle).toBoolean()
+        val isFirstRun = LibGleanFFI.INSTANCE.glean_is_first_run().toBoolean()
         if (isFirstRun) {
             initializeCoreMetrics(applicationContext)
         }
@@ -174,9 +172,7 @@ open class GleanInternalAPI internal constructor () {
         // Deal with any pending events so we can start recording new ones
         @Suppress("EXPERIMENTAL_API_USAGE")
         Dispatchers.API.executeTask {
-            val pingSubmitted = LibGleanFFI.INSTANCE.glean_on_ready_to_submit_pings(
-                this@GleanInternalAPI.handle
-            ).toBoolean()
+            val pingSubmitted = LibGleanFFI.INSTANCE.glean_on_ready_to_submit_pings().toBoolean()
             if (pingSubmitted) {
                 PingUploadWorker.enqueueWorker(applicationContext)
             }
@@ -194,7 +190,7 @@ open class GleanInternalAPI internal constructor () {
         if (!isFirstRun) {
             @Suppress("EXPERIMENTAL_API_USAGE")
             Dispatchers.API.executeTask {
-                LibGleanFFI.INSTANCE.glean_clear_application_lifetime_metrics(handle)
+                LibGleanFFI.INSTANCE.glean_clear_application_lifetime_metrics()
                 initializeCoreMetrics(applicationContext)
             }
         }
@@ -220,7 +216,7 @@ open class GleanInternalAPI internal constructor () {
      * Returns true if the Glean SDK has been initialized.
      */
     internal fun isInitialized(): Boolean {
-        return handle != 0L
+        return initialized
     }
 
     /**
@@ -259,7 +255,7 @@ open class GleanInternalAPI internal constructor () {
                 // therefore need to obtain the lock from the PingUploader so that
                 // iterating over and deleting the pings doesn't happen at the same time.
                 synchronized(PingUploadWorker.pingQueueLock) {
-                    LibGleanFFI.INSTANCE.glean_set_upload_enabled(handle, enabled.toByte())
+                    LibGleanFFI.INSTANCE.glean_set_upload_enabled(enabled.toByte())
 
                     // Cancel any pending workers here so that we don't accidentally upload or
                     // collect data after the upload has been disabled.
@@ -290,7 +286,7 @@ open class GleanInternalAPI internal constructor () {
      */
     fun getUploadEnabled(): Boolean {
         if (isInitialized()) {
-            return LibGleanFFI.INSTANCE.glean_is_upload_enabled(handle).toBoolean()
+            return LibGleanFFI.INSTANCE.glean_is_upload_enabled().toBoolean()
         } else {
             return uploadEnabled
         }
@@ -332,7 +328,6 @@ open class GleanInternalAPI internal constructor () {
         @Suppress("EXPERIMENTAL_API_USAGE")
         Dispatchers.API.launch {
             LibGleanFFI.INSTANCE.glean_set_experiment_active(
-                handle,
                 experimentId,
                 branch,
                 keys,
@@ -352,7 +347,7 @@ open class GleanInternalAPI internal constructor () {
         // initialized, it doesn't get ignored and will be replayed after init.
         @Suppress("EXPERIMENTAL_API_USAGE")
         Dispatchers.API.launch {
-            LibGleanFFI.INSTANCE.glean_set_experiment_inactive(handle, experimentId)
+            LibGleanFFI.INSTANCE.glean_set_experiment_inactive(experimentId)
         }
     }
 
@@ -367,7 +362,7 @@ open class GleanInternalAPI internal constructor () {
         @Suppress("EXPERIMENTAL_API_USAGE")
         Dispatchers.API.assertInTestingMode()
 
-        return LibGleanFFI.INSTANCE.glean_experiment_test_is_active(handle, experimentId).toBoolean()
+        return LibGleanFFI.INSTANCE.glean_experiment_test_is_active(experimentId).toBoolean()
     }
 
     /**
@@ -398,7 +393,6 @@ open class GleanInternalAPI internal constructor () {
         Dispatchers.API.assertInTestingMode()
 
         val ptr = LibGleanFFI.INSTANCE.glean_experiment_test_get_data(
-            handle,
             experimentId
         )!!
 
@@ -473,7 +467,7 @@ open class GleanInternalAPI internal constructor () {
      */
     @VisibleForTesting(otherwise = VisibleForTesting.NONE)
     internal fun testCollect(ping: PingType): String? {
-        return LibGleanFFI.INSTANCE.glean_ping_collect(handle, ping.handle)?.getAndConsumeRustString()
+        return LibGleanFFI.INSTANCE.glean_ping_collect(ping.handle)?.getAndConsumeRustString()
     }
 
     /**
@@ -553,7 +547,6 @@ open class GleanInternalAPI internal constructor () {
         val pingArray = StringArray(pingNames.toTypedArray(), "utf-8")
         val pingArrayLen = pingNames.size
         val submittedPing = LibGleanFFI.INSTANCE.glean_submit_pings_by_name(
-            handle,
             pingArray,
             pingArrayLen
         ).toBoolean()
@@ -594,7 +587,7 @@ open class GleanInternalAPI internal constructor () {
 
         if (isInitialized() && clearStores) {
             // Clear all the stored data.
-            LibGleanFFI.INSTANCE.glean_test_clear_all_stores(handle)
+            LibGleanFFI.INSTANCE.glean_test_clear_all_stores()
         }
 
         isMainProcess = null
@@ -633,12 +626,12 @@ open class GleanInternalAPI internal constructor () {
     @VisibleForTesting(otherwise = VisibleForTesting.NONE)
     internal fun testDestroyGleanHandle() {
         if (!isInitialized()) {
-            // We don't need to destroy the Glean handle: it wasn't initialized.
+            // We don't need to destroy Glean: it wasn't initialized.
             return
         }
 
-        LibGleanFFI.INSTANCE.glean_destroy_glean(handle)
-        handle = 0L
+        LibGleanFFI.INSTANCE.glean_destroy_glean()
+        initialized = false
     }
 
     /**
@@ -648,7 +641,6 @@ open class GleanInternalAPI internal constructor () {
     internal fun registerPingType(pingType: PingType) {
         if (this.isInitialized()) {
             LibGleanFFI.INSTANCE.glean_register_ping_type(
-                handle,
                 pingType.handle
             )
         }
@@ -668,7 +660,7 @@ open class GleanInternalAPI internal constructor () {
      */
     @VisibleForTesting(otherwise = VisibleForTesting.NONE)
     internal fun testHasPingType(pingName: String): Boolean {
-        return LibGleanFFI.INSTANCE.glean_test_has_ping_type(handle, pingName).toBoolean()
+        return LibGleanFFI.INSTANCE.glean_test_has_ping_type(pingName).toBoolean()
     }
 
     /**