+++ glean-core/android/src/main/java/mozilla/telemetry/glean/private/CounterMetricType.kt
@@ -7,7 +7,6 @@ package mozilla.telemetry.glean.private
 import androidx.annotation.VisibleForTesting
 import com.sun.jna.StringArray
 import mozilla.telemetry.glean.Dispatchers
-import mozilla.telemetry.glean.Glean
 import mozilla.telemetry.glean.rust.LibGleanFFI
 import mozilla.telemetry.glean.rust.toBoolean
 import mozilla.telemetry.glean.rust.toByte
@@ -73,7 +72,6 @@ class CounterMetricType internal constructor(
         @Suppress("EXPERIMENTAL_API_USAGE")
         Dispatchers.API.launch {
             LibGleanFFI.INSTANCE.glean_counter_add(
-                Glean.handle,
                 this@CounterMetricType.handle,
                 amount)
         }
@@ -92,7 +90,6 @@ class CounterMetricType internal constructor(
         }
 
         LibGleanFFI.INSTANCE.glean_counter_add(
-            Glean.handle,
             this@CounterMetricType.handle,
             amount
         )
@@ -113,7 +110,7 @@ class CounterMetricType internal constructor(
         @Suppress("EXPERIMENTAL_API_USAGE")
         Dispatchers.API.assertInTestingMode()
 
-        val res = LibGleanFFI.INSTANCE.glean_counter_test_has_value(Glean.handle, this.handle, pingName)
+        val res = LibGleanFFI.INSTANCE.glean_counter_test_has_value(this.handle, pingName)
         return res.toBoolean()
     }
 
@@ -135,7 +132,7 @@ class CounterMetricType internal constructor(
         if (!testHasValue(pingName)) {
             throw NullPointerException()
         }
-        return LibGleanFFI.INSTANCE.glean_counter_test_get_value(Glean.handle, this.handle, pingName)
+        return LibGleanFFI.INSTANCE.glean_counter_test_get_value(this.handle, pingName)
     }
 
     /**
@@ -153,7 +150,7 @@ class CounterMetricType internal constructor(
         Dispatchers.API.assertInTestingMode()
 
         return LibGleanFFI.INSTANCE.glean_counter_test_get_num_recorded_errors(
-            Glean.handle, this.handle, errorType.ordinal, pingName
+            this.handle, errorType.ordinal, pingName
         )
     }
 }