+++ glean-core/ffi/src/string.rs
@@ -7,7 +7,8 @@ use std::os::raw::c_char;
 use ffi_support::FfiStr;
 
 use crate::{
-    define_metric, ffi_string_ext::FallibleToString, handlemap_ext::HandleMapExtension, GLEAN,
+    define_metric, ffi_string_ext::FallibleToString, handlemap_ext::HandleMapExtension,
+    with_glean_value,
 };
 
 define_metric!(StringMetric => STRING_METRICS {
@@ -17,40 +18,34 @@ define_metric!(StringMetric => STRING_METRICS {
 });
 
 #[no_mangle]
-pub extern "C" fn glean_string_set(glean_handle: u64, metric_id: u64, value: FfiStr) {
-    GLEAN.call_infallible(glean_handle, |glean| {
+pub extern "C" fn glean_string_set(metric_id: u64, value: FfiStr) {
+    with_glean_value(|glean| {
         STRING_METRICS.call_with_log(metric_id, |metric| {
             let value = value.to_string_fallible()?;
-            metric.set(glean, value);
+            metric.set(&glean, value);
             Ok(())
         })
     })
 }
 
 #[no_mangle]
-pub extern "C" fn glean_string_test_has_value(
-    glean_handle: u64,
-    metric_id: u64,
-    storage_name: FfiStr,
-) -> u8 {
-    GLEAN.call_infallible(glean_handle, |glean| {
+pub extern "C" fn glean_string_test_has_value(metric_id: u64, storage_name: FfiStr) -> u8 {
+    with_glean_value(|glean| {
         STRING_METRICS.call_infallible(metric_id, |metric| {
             metric
-                .test_get_value(glean, storage_name.as_str())
+                .test_get_value(&glean, storage_name.as_str())
                 .is_some()
         })
     })
 }
 
 #[no_mangle]
-pub extern "C" fn glean_string_test_get_value(
-    glean_handle: u64,
-    metric_id: u64,
-    storage_name: FfiStr,
-) -> *mut c_char {
-    GLEAN.call_infallible(glean_handle, |glean| {
+pub extern "C" fn glean_string_test_get_value(metric_id: u64, storage_name: FfiStr) -> *mut c_char {
+    with_glean_value(|glean| {
         STRING_METRICS.call_infallible(metric_id, |metric| {
-            metric.test_get_value(glean, storage_name.as_str()).unwrap()
+            metric
+                .test_get_value(&glean, storage_name.as_str())
+                .unwrap()
         })
     })
 }