+++ glean-core/ffi/src/string_list.rs
@@ -8,7 +8,7 @@ use ffi_support::FfiStr;
 
 use crate::{
     define_metric, ffi_string_ext::FallibleToString, from_raw_string_array,
-    handlemap_ext::HandleMapExtension, RawStringArray, GLEAN,
+    handlemap_ext::HandleMapExtension, with_glean_value, RawStringArray,
 };
 
 define_metric!(StringListMetric => STRING_LIST_METRICS {
@@ -18,8 +18,8 @@ define_metric!(StringListMetric => STRING_LIST_METRICS {
 });
 
 #[no_mangle]
-pub extern "C" fn glean_string_list_add(glean_handle: u64, metric_id: u64, value: FfiStr) {
-    GLEAN.call_infallible(glean_handle, |glean| {
+pub extern "C" fn glean_string_list_add(metric_id: u64, value: FfiStr) {
+    with_glean_value(|glean| {
         STRING_LIST_METRICS.call_with_log(metric_id, |metric| {
             let value = value.to_string_fallible()?;
             metric.add(glean, value);
@@ -29,13 +29,8 @@ pub extern "C" fn glean_string_list_add(glean_handle: u64, metric_id: u64, value
 }
 
 #[no_mangle]
-pub extern "C" fn glean_string_list_set(
-    glean_handle: u64,
-    metric_id: u64,
-    values: RawStringArray,
-    values_len: i32,
-) {
-    GLEAN.call_infallible(glean_handle, |glean| {
+pub extern "C" fn glean_string_list_set(metric_id: u64, values: RawStringArray, values_len: i32) {
+    with_glean_value(|glean| {
         STRING_LIST_METRICS.call_with_log(metric_id, |metric| {
             let values = from_raw_string_array(values, values_len)?;
             metric.set(glean, values);
@@ -45,12 +40,8 @@ pub extern "C" fn glean_string_list_set(
 }
 
 #[no_mangle]
-pub extern "C" fn glean_string_list_test_has_value(
-    glean_handle: u64,
-    metric_id: u64,
-    storage_name: FfiStr,
-) -> u8 {
-    GLEAN.call_infallible(glean_handle, |glean| {
+pub extern "C" fn glean_string_list_test_has_value(metric_id: u64, storage_name: FfiStr) -> u8 {
+    with_glean_value(|glean| {
         STRING_LIST_METRICS.call_infallible(metric_id, |metric| {
             metric
                 .test_get_value(glean, storage_name.as_str())
@@ -61,11 +52,10 @@ pub extern "C" fn glean_string_list_test_has_value(
 
 #[no_mangle]
 pub extern "C" fn glean_string_list_test_get_value_as_json_string(
-    glean_handle: u64,
     metric_id: u64,
     storage_name: FfiStr,
 ) -> *mut c_char {
-    GLEAN.call_infallible(glean_handle, |glean| {
+    with_glean_value(|glean| {
         STRING_LIST_METRICS.call_infallible(metric_id, |metric| {
             metric
                 .test_get_value_as_json_string(glean, storage_name.as_str())