+++ glean-core/ios/GleanTests/Metrics/PingTests.swift
@@ -118,7 +118,7 @@ class PingTests: XCTestCase {
 
         Glean.shared.submitPingsByName(pingNames: ["unknown"])
 
-        /// We wait for a timeout to happen, as we don't expect any data to be sent.
+        // We wait for a timeout to happen, as we don't expect any data to be sent.
         waitForExpectations(timeout: 5.0) { _ in
             XCTAssert(true, "Test didn't time out when it should")
         }