+++ glean-core/python/glean/metrics/event.py
@@ -7,7 +7,6 @@ import json
 from typing import Dict, List, Optional
 
 
-from ..glean import Glean
 from .. import _ffi
 from .._dispatcher import Dispatcher
 from ..testing import ErrorType
@@ -140,7 +139,6 @@ class EventMetricType:
                 nextra = len(extra)
 
             _ffi.lib.glean_event_record(
-                Glean._handle,
                 self._handle,
                 timestamp,
                 _ffi.ffi_encode_vec_int32(keys),
@@ -165,7 +163,7 @@ class EventMetricType:
 
         return bool(
             _ffi.lib.glean_event_test_has_value(
-                Glean._handle, self._handle, _ffi.ffi_encode_string(ping_name)
+                self._handle, _ffi.ffi_encode_string(ping_name)
             )
         )
 
@@ -190,7 +188,7 @@ class EventMetricType:
 
         json_string = _ffi.ffi_decode_string(
             _ffi.lib.glean_event_test_get_value_as_json_string(
-                Glean._handle, self._handle, _ffi.ffi_encode_string(ping_name)
+                self._handle, _ffi.ffi_encode_string(ping_name)
             )
         )
 
@@ -217,10 +215,7 @@ class EventMetricType:
             ping_name = self._send_in_pings[0]
 
         return _ffi.lib.glean_event_test_get_num_recorded_errors(
-            Glean._handle,
-            self._handle,
-            error_type.value,
-            _ffi.ffi_encode_string(ping_name),
+            self._handle, error_type.value, _ffi.ffi_encode_string(ping_name),
         )